requirements.txt
pytest==8.4.2
selenium==4.38.0
allure-pytest==2.15.0
faker==37.12.0